Abstract Operation functionality to it's own class
