+++ src/Services/BingAds/Operations/AdGroupOperations.php
@@ -2,36 +2,19 @@
 
 namespace LaravelAds\Services\BingAds\Operations;
 
-use LaravelAds\Services\BingAds\Operations\AdGroup;
 use LaravelAds\Services\BingAds\Service;
+use Microsoft\BingAds\Auth\ServiceClient;
 
-use Microsoft\BingAds\V13\CampaignManagement\GetAdGroupsByIdsRequest;
+use Microsoft\BingAds\Auth\ServiceClientType;
+use LaravelAds\Services\BingAds\Operations\AdGroup;
+use LaravelAds\Services\BingAds\Operations\Operation;
+
 use Microsoft\BingAds\V13\CampaignManagement\UpdateAdGroupsRequest;
 use Microsoft\BingAds\V13\CampaignManagement\AdGroup as AdGroupProxy;
+use Microsoft\BingAds\V13\CampaignManagement\GetAdGroupsByIdsRequest;
 
-use Microsoft\BingAds\Auth\ServiceClient;
-use Microsoft\BingAds\Auth\ServiceClientType;
-
-class AdGroupOperations
+class AdGroupOperations extends Operation
 {
-    /**
-     * $service
-     *
-     */
-    protected $service = null;
-
-    /**
-     * $campaignRequest
-     *
-     */
-    protected $request = null;
-
-    /**
-     * $campaignResponse
-     *
-     */
-    protected $response = null;
-
     /**
      * __construct()
      *
@@ -41,49 +24,6 @@ class AdGroupOperations
         $this->service = $service;
 
         $this->request = new AdGroupProxy();
-    }
-
-    /**
-     * request()
-     *
-     */
-    public function request()
-    {
-        return $this->request;
-    }
-
-    /**
-     * response()
-     *
-     */
-    public function response()
-    {
-        return $this->response;
-    }
-
-    /**
-     * set()
-     *
-     */
-    public function set($adGroup)
-    {
-        $this->response = $adGroup;
-
-        // set up our request if we have not done this yet
-        $this->request()->Id = $adGroup->Id;
-
-        return $this;
-    }
-
-    /**
-     * get()
-     *
-     */
-    public function get()
-    {
-        $this->set($this->sendRequest());
-
-        return $this;
     }
 
     /**
